Move plugin startup sync into core-plugins
Move curated plugin startup sync and installed marketplace root discovery into codex-core-plugins using narrow codex_home and ConfigLayerStack inputs. Leave codex-core with only the remote startup sync glue that still needs Config, AuthManager, and PluginsManager.
