--- codex-rs/core/src/plugins/installed_marketplaces.rs
@@ -1,55 +1,0 @@
-use crate::config::Config;
-use codex_core_plugins::installed_marketplaces::marketplace_install_root;
-use codex_core_plugins::installed_marketplaces::resolve_configured_marketplace_root;
-use codex_core_plugins::marketplace::find_marketplace_manifest_path;
-use codex_utils_absolute_path::AbsolutePathBuf;
-use std::path::Path;
-use tracing::warn;
-
-use super::validate_plugin_segment;
-
-pub(crate) fn installed_marketplace_roots_from_config(
-    config: &Config,
-    codex_home: &Path,
-) -> Vec<AbsolutePathBuf> {
-    let Some(user_layer) = config.config_layer_stack.get_user_layer() else {
-        return Vec::new();
-    };
-    let Some(marketplaces_value) = user_layer.config.get("marketplaces") else {
-        return Vec::new();
-    };
-    let Some(marketplaces) = marketplaces_value.as_table() else {
-        warn!("invalid marketplaces config: expected table");
-        return Vec::new();
-    };
-    let default_install_root = marketplace_install_root(codex_home);
-    let mut roots = marketplaces
-        .iter()
-        .filter_map(|(marketplace_name, marketplace)| {
-            if !marketplace.is_table() {
-                warn!(
-                    marketplace_name,
-                    "ignoring invalid configured marketplace entry"
-                );
-                return None;
-            }
-            if let Err(err) = validate_plugin_segment(marketplace_name, "marketplace name") {
-                warn!(
-                    marketplace_name,
-                    error = %err,
-                    "ignoring invalid configured marketplace name"
-                );
-                return None;
-            }
-            let path = resolve_configured_marketplace_root(
-                marketplace_name,
-                marketplace,
-                &default_install_root,
-            )?;
-            find_marketplace_manifest_path(&path).map(|_| path)
-        })
-        .filter_map(|path| AbsolutePathBuf::try_from(path).ok())
-        .collect::<Vec<_>>();
-    roots.sort_unstable_by(|left, right| left.as_path().cmp(right.as_path()));
-    roots
-}
